item/actions/effects: merge ITEM_ACTION_FLOOD handler
