--- src/tr1/game/item_actions/flood.h
@@ -1,5 +1,0 @@
-#pragma once
-
-#include <libtrx/game/items.h>
-
-void ItemAction_Flood(ITEM *item);
