+++ src/tr1/meson.build
@@ -118,7 +118,6 @@ sources = [
   'game/item_actions/finish_level.c',
   'game/item_actions/flicker.c',
   'game/item_actions/flipmap.c',
-  'game/item_actions/flood.c',
   'game/item_actions/floor_shake.c',
   'game/item_actions/powerup.c',
   'game/item_actions/raising_block.c',
